refactor: 얕은 비교, 깊은 비교 함수 공통 로직 baseEquals로 묶음
