+++ src/@lib/equalities/shallowEquals.ts
@@ -1,32 +1,5 @@
+import { baseEquals } from "./baseEquals";
+
 export function shallowEquals<T>(objA: T, objB: T): boolean {
-  if (Object.is(objA, objB)) {
-    return true;
-  }
-
-  if (
-    typeof objA !== "object" ||
-    objA === null ||
-    typeof objB !== "object" ||
-    objB === null
-  ) {
-    return false;
-  }
-
-  const objAKeys = Object.keys(objA);
-  const objBKeys = Object.keys(objB);
-
-  if (objAKeys.length !== objBKeys.length) {
-    return false;
-  }
-
-  for (const key of objAKeys) {
-    if (
-      !Reflect.has(objB, key) ||
-      !Object.is(Reflect.get(objA, key), Reflect.get(objB, key))
-    ) {
-      return false;
-    }
-  }
-
-  return true;
+  return baseEquals(objA, objB, Object.is);
 }
